backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.8.0
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.1.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
msgpack==1.1.2
packaging==25.0
pluggy==1.6.0
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==9.0.2
pytest-django==4.9.0
pytest-cov==6.0.0
python-environ==0.4.54
redis==7.1.0
service-identity==24.2.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.11.0
zope.interface==8.2
